api/requirements.txt
azure-functions
requests
opencensus-ext-azure
opencensus-ext-requests